refactor: move utils modules to utils/ package
Phase 1 of incremental reorganization (issue #625):
- entity_validator.py -> utils/validation.py
- cost_tracker.py -> utils/costs.py
- Update all imports to use new module paths
- Move corresponding test files to tests/utils/

All 700 tests pass.
