+++ custom_components/localshift/state_machine.py
@@ -33,8 +33,8 @@ if TYPE_CHECKING:
     from .battery_controller import BatteryController
     from .computation_engine import ComputationEngine
     from .computation_engine_lib.decision_outcome_tracker import DecisionOutcomeTracker
-    from .entity_validator import EntityValidator
     from .notification_service import NotificationService
+    from .utils.validation import EntityValidator
 
 
 _LOGGER = logging.getLogger(__name__)
