+++ tests/test_state_machine.py
@@ -86,7 +86,7 @@ def mock_get_option():
 @pytest.fixture
 def mock_entity_validator():
     """Create a mock EntityValidator."""
-    from custom_components.localshift.entity_validator import IntegrationStatus
+    from custom_components.localshift.utils.validation import IntegrationStatus
 
     validator = MagicMock()
     validator.should_allow_automation = MagicMock(return_value=True)
