+++ tests/test_state_reader.py
@@ -45,7 +45,7 @@ def mock_entry():
 @pytest.fixture
 def mock_entity_validator():
     """Create a mock EntityValidator for testing."""
-    from custom_components.localshift.entity_validator import IntegrationStatus
+    from custom_components.localshift.utils.validation import IntegrationStatus
 
     validator = MagicMock()
     validator.should_allow_automation = MagicMock(return_value=True)
